Revert "UltraRPS (complete)"
This reverts commit f92ff890cd691fa41ebdbfefe0316bbabfdabfd7.
