+++ src/main/resources/config.json
@@ -1,6 +1,6 @@
 {
 
    "channels": ["rachel"],
-   "token": "NzU2OTk2OTIxNjIwNTYyMDcz.X2Z-Ug.K06Se_pF2KHtIjy2e0ot0nGLN90"
+   "token": "NzU2OTk2OTIxNjIwNTYyMDcz.X2Z-Ug.FN-no6OCRvY47vvFL9bA911UfnI"
 
 }
